Bug fixes:
Issue with running out of food in later game.
Issue with laser turrets.
Issue with getting locked in endless battle (F9)
Fixed logic of final battle:

So, the final battle was set to go off when you *built* the launchpad,
not when you actually launched. I thought this was the desired behavior
and I went with it. Now if you lose the battle, your launchpad is
destroyed, and if you win, you can beat the game at any time by choosing
Initiate Launch (no battle happens). (Also, if you lose, you can beat
the game during the 10-second window before the launchpad's
demolition comes back from the server.) Overall I think this behavior
is fine, and it's good to go.

Also I have it where aliens that would have attacked the HQ now attack
the launchpad. Some aliens will randomly attack the HQ as well, and
they could even bring it down to 0 health, but the battle will continue.
I think this behavior is fine as well.

Also added controls to readme